feat(web): update middleware
